hand-eye-calibration-main/utils/arm_rw.py
from typing import Tuple, Optional
from xarm.wrapper import XArmAPI

class XArm7RW():
    """Implementation of BaseRobotArm for xArm7 real-world control."""
    def __init__(self, ip: str):
        """Initialize xArm7 controller.
        
        Args:
            ip: IP address of the xArm controller
        """
        self.arm = XArmAPI(ip)

    def get_joint_values(self, is_radian: bool = True) -> Tuple[float, ...]:
        """Get current joint angles of xArm7.
        
        Args:
            is_radian: Whether to return values in radians (True) or degrees (False)
            
        Returns:
            Tuple of 7 joint angles
        """
        state, joint_values = self.arm.get_servo_angle(is_radian=is_radian)
        if state != 0:
            raise RuntimeError(f"Failed to get joint values (error code: {state})")
        
        return joint_values[:7]  # Return first 7 joints
    
    def get_ee_pose(self, is_radian: bool = False) -> Tuple[float, ...]:
        """Get end-effector pose of xArm7.
        
        Args:
            is_radian: Whether to return orientation in radians (True) or degrees (False)
        Returns:
            Tuple of (x, y, z, roll, pitch, yaw) in mm and radians/degrees
        """
        code, pose = self.arm.get_position(is_radian=is_radian)
        if code != 0:
            raise RuntimeError(f"Failed to get end-effector pose (error code: {code})")
        return pose
    
"""Example of creating a new robot type:"""
class MyCustomArm():
    """Custom robot arm implementation."""
    def __init__(self, ip: str):
        # Initialization of the robot arm
        pass

    def get_joint_values(self, is_radian: bool = True) -> Tuple[float, ...]:
        # Custom implementation for getting joint values
        pass

    def get_ee_pose(self, is_radian: bool = False) -> Tuple[float, ...]:
        # Custom implementation for getting end-effector pose
        pass